requirements.txt
# FastAPI
fastapi
pydantic
uvicorn
asgi-logger
gunicorn
motor

python-jose[cryptography]
httpx
